PhotoLogicManager.cs finished.
Player can now take a photo, then edit the photo to be used for AR minigames.

Still needs to be connected to actual models to be displayed on the phone, and for further iterations on the AR mechanics to see if it will recognize the photo.

Notion: This update is for 1st prototype> Game Mechanics> AR Recognition Mechanics> Scannable Drawings.